refactor(desktop): organize workspace command palette hooks
