+++ desktop/src/components/workspace/shell/command-palette/WorkspaceCommandPalette.tsx
@@ -11,7 +11,7 @@ import { FileTreeEntryIcon } from "@/components/ui/file-icons";
 import {
   CommandPaletteGlyph,
 } from "@/components/ui/icons";
-import { useWorkspaceCommandPalette } from "@/hooks/workspaces/use-workspace-command-palette";
+import { useWorkspaceCommandPalette } from "@/hooks/workspaces/facade/use-workspace-command-palette";
 import type {
   CommandPaletteEntry,
   CommandPaletteIconId,
