+++ desktop/src/hooks/workspaces/derived/use-workspace-command-palette-open-files.ts
@@ -14,6 +14,7 @@ export interface CommandPaletteOpenFileEntry {
   isActive: boolean;
 }
 
+// Derives open file entries for command-palette rendering from viewer tab state.
 export function useWorkspaceCommandPaletteOpenFiles(
   selectedWorkspaceId: string | null,
 ): CommandPaletteOpenFileEntry[] {
